+++ src/docs/DocsNavigation/links.ts
@@ -47,7 +47,7 @@ export const menuNavLinks: any = [
 			{ href: '/components/data-tables', label: 'Data Tables' },
 			{ href: '/components/dividers', label: 'Dividers' },
 			{ href: '/components/gradient-headings', label: 'Gradient Headings' },
-			{ href: '/components/lists', label: 'Listboxes', badge: 'Updated' },
+			{ href: '/components/listboxes', label: 'Listboxes', badge: 'Updated' },
 			{ href: '/components/menus', label: 'Menus' },
 			{ href: '/components/paginators', label: 'Paginators' },
 			{ href: '/components/placeholders', label: 'Placeholders' },
