+++ src/lib/components/ListBox/ListBoxItem.svelte
