+++ src/lib/index.ts
@@ -15,8 +15,8 @@ export { default as Button } from './components/Button/Button.svelte';
 export { default as Card } from './components/Card/Card.svelte';
 export { default as Divider } from './components/Divider/Divider.svelte';
 export { default as GradientHeading } from './components/GradientHeading/GradientHeading.svelte';
-export { default as List } from './components/List/List.svelte';
-export { default as ListItem } from './components/List/ListItem.svelte';
+export { default as ListBox } from './components/ListBox/ListBox.svelte';
+export { default as ListBoxItem } from './components/ListBox/ListBoxItem.svelte';
 export { default as LogoCloud } from './components/LogoCloud/LogoCloud.svelte';
 export { default as Logo } from './components/LogoCloud/Logo.svelte';
 export { default as Menu } from './components/Menu/Menu.svelte';
